backend/category_offers/requirements.txt
django > 4
djangorestframework
django-extensions
django-cors-headers
rich
factory_boy